# docs/contract/openapi-m2-7.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI 研究者助手 · M2-7 累积接口契约冻结（M1 + M2-1 + M2-2 + M2-4 + M2-5 + M2-7）",
    "version": "0.1.0",
    "description": "本 schema 为截至 M2-7（数据点级溯源）后端冻结契约的累积快照，在 M2-5 全部端点之上新增一个报告只读端点：\nGET /reports/{run_id}/citations 返回报告级信源索引（按 marker [N] 升序、报告级去重），元素为 ReportCitationItem：内联三字段 evidence_id/marker/snippet 之外补七项回溯展示字段 url/title/domain/source_type/source_level/credibility/published_at（marker/snippet 与该七项合计九展示字段，published_at 可空）。\n既有 GET /reports/{run_id} 与 GET /runs/{run_id}/report 响应在原八字段之上增 outline（sec-overview/sec-findings/sec-disputes/sec-limitations 语义章节，分歧/局限章按需出现）与 blocks（conclusion/evidence/dispute/limitation 四类，conclusion/dispute 带 claim_id 与 confidence，dispute 带 conflict_id，block 内联 citations 为 {evidence_id, marker, snippet}）；draft/历史报告两数组为空。\n终稿 status=final，含数字/事实声明的论断 100% 绑定信源（缺源降级 inferred，无来源数字块经一次自修复仍不达标则剔除）；同一证据跨块共享 marker，未被引用证据不占号。WebSocket 帧在本批零变化。"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "发起研究运行",
        "description": "创建研究运行并后台调度 executor。\n\n校验项目归属：``project_id`` 必须属于当前用户团队且 ``status='active'``。",
        "operationId": "create_run_api_v1_runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "查询研究运行列表（创建者维度）",
        "description": "分页返回当前用户创建的研究运行（created_at 倒序）。\n\n- ``status``：六态单值过滤，非法枚举由 FastAPI 判 422；\n- ``project_id``：再过滤且校验团队归属，非本团队项目 404。",
        "operationId": "list_runs_api_v1_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "running",
                    "paused",
                    "succeeded",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 26
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/pause": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "暂停研究运行（软暂停）",
        "description": "软暂停运行中的研究。\n\n- 仅 ``running`` 且存在在途执行协程时可暂停；已暂停 409 RUN_ALREADY_PAUSED，\n  其他终态/孤儿 running 409 RUN_NOT_PAUSABLE；\n- 暂停为协作式：响应返回后协程在最近 super-step 边界落 paused 并推\n  ``run.finished(status=paused)``，恢复走 POST /resume（M2-5 T2）。",
        "operationId": "pause_run_api_v1_runs__run_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "取消研究运行（硬中断）",
        "description": "硬取消研究；对 succeeded/failed/cancelled 幂等返回当前状态。\n\n``keep_partial=true``（默认）时若已有报告草稿则保留并在响应带回 id。",
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "恢复研究运行（人类输入统一入口）",
        "description": "从暂停/澄清/裁决/成本挂起点恢复。\n\n空 body 或 ``{human_input:{kind:\"proceed\"}}`` 为纯继续；澄清挂起必须带\n非空 answers；载荷校验与 409/422 语义见 M2-5 方案 §5.3。",
        "operationId": "resume_run_api_v1_runs__run_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/intervene": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "运行中主动介入（追加追问/剔除证据）",
        "description": "running 中的受控介入；动作语义与可用阶段见 M2-5 方案 §5.5。",
        "operationId": "intervene_run_api_v1_runs__run_id__intervene_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "查询研究运行状态",
        "description": "查询单个研究运行的当前状态。\n\npaused@clarify 时附加 interrupt 澄清上下文（M2-5 §5.4，供刷新恢复澄清卡）。",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/report": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "查询研究运行的报告",
        "description": "查询研究运行产出的报告（Markdown 正文 + M2-7 结构化 outline/blocks 超集）。",
        "operationId": "get_run_report_api_v1_runs__run_id__report_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/stages": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "查询运行阶段时间线",
        "description": "固定六阶段顺序返回；历史无阶段行的 run 返回空数组。",
        "operationId": "list_run_stages_api_v1_runs__run_id__stages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StageResponse"
                  },
                  "title": "Response List Run Stages Api V1 Runs  Run Id  Stages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/sub-questions": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "查询运行的子问题列表",
        "description": "按 created_at 升序返回 run 下全部子问题。",
        "operationId": "list_run_sub_questions_api_v1_runs__run_id__sub_questions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubQuestionResponse"
                  },
                  "title": "Response List Run Sub Questions Api V1 Runs  Run Id  Sub Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{run_id}/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "查询单条证据详情（全文懒加载）",
        "description": "单条证据详情；归属不符/不存在统一 404。",
        "operationId": "get_run_evidence_api_v1_runs__run_id__evidence__evidence_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "maxLength": 26,
            "minLength": 1,
            "title": "Project Id"
          },
          "question": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 5,
            "title": "Question"
          },
          "tier": {
            "type": "string",
            "enum": [
              "quick",
              "standard",
              "deep",
              "extreme"
            ],
            "title": "Tier",
            "default": "standard"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "question"
        ],
        "title": "CreateRunRequest",
        "description": "``POST /runs`` 请求体。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PauseRunRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "PauseRunRequest",
        "description": "``POST /runs/{id}/pause`` 请求体（契约草案 §6.1）。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT Bearer token",
        "scheme": "bearer"
      }
    }
  }
}